third_party/blink/web_tests/wpt_internal/fenced_frame/disallowed-navigations-dangling-markup.https.html

<!DOCTYPE html>
<title>Fenced frame disallowed navigations with potentially-dangling markup</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/utils.js"></script>
<script src="/fetch/private-network-access/resources/support.sub.js"></script>

<body>

<script>
// These are used in tests that rely on URLs containing dangling markup. See
// https://github.com/whatwg/fetch/pull/519.
const kDanglingMarkupSubstrings = [
  "blo\nck<ed",
  "blo\rck<ed",
  "blo\tck<ed",
  "blo<ck\ned",
  "blo<ck\red",
  "blo<ck\ted",
];

function getTimeoutPromise(t) {
  return new Promise(resolve =>
      t.step_timeout(() => resolve("NOT LOADED"), 1500));
}

// The following tests ensure that an embedder cannot navigate a fenced frame
// (with different modes) to https URLs with dangling markup
for (const mode of ['opaque-ads', 'default']) {
  // These tests assert that fenced frames cannot be navigated to HTTPs URLs
  // with dangling markup.
  for (substring of kDanglingMarkupSubstrings) {
    promise_test(async t => {
      const key = token();
      let url_string = generateURL("resources/embeddee.html?blocked", [key]).toString();
      url_string = url_string.replace("blocked", substring);
      const fencedframe = attachFencedFrame(url_string, /*mode=*/mode);
      const loaded_promise = nextValueFromServer(key);
      const result = await Promise.any([loaded_promise, getTimeoutPromise(t)]);
      assert_equals(result, "NOT LOADED");
    }, `fenced frame mode=${mode} dangling-markup URL with '${substring}'`);
  }
} // end for.

// These tests assert that fenced frames cannot be navigated to a urn:uuid URL
// that represents an HTTPS URLs with dangling markup.
for (substring of kDanglingMarkupSubstrings) {
  promise_test(async t => {
    const key = token();

    // Copied from from `generateURN()`, since we have to modify the final URL
    // that goes into `selectURL`.
    try {
      await sharedStorage.worklet.addModule(
        "/wpt_internal/shared_storage/resources/simple-module.js");
    } catch (e) {
      // See documentation in `generateURN()`.
    }

    let url_string = generateURL("resources/report-url.html?blocked", [key]).toString();
    url_string = url_string.replace("blocked", substring);

    const urn = await sharedStorage.selectURL(
      "test-url-selection-operation", [{url: url_string}], {data: {'mockResult': 0}}
    );

    const fencedframe = attachFencedFrame(urn, /*mode=*/'opaque-ads');
    const loaded_promise = nextValueFromServer(key);
    const result = await Promise.any([loaded_promise, getTimeoutPromise(t)]);
    assert_equals(result, "NOT LOADED");
  }, `fenced frame opaque URN => https: URL with dangling markup '${substring}'`);
}

</script>

</body>
