third_party/blink/web_tests/platform/generic/virtual/fenced-frame-shadow-dom/wpt_internal/fenced_frame/disallowed-navigations-dangling-markup.https-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = ResizeObserver loop limit exceeded
PASS fenced frame mode=opaque-ads dangling-markup URL with 'blo
ck<ed'
PASS fenced frame mode=opaque-ads dangling-markup URL with 'blo\rck<ed'
PASS fenced frame mode=opaque-ads dangling-markup URL with 'blo	ck<ed'
PASS fenced frame mode=opaque-ads dangling-markup URL with 'blo<ck
ed'
PASS fenced frame mode=opaque-ads dangling-markup URL with 'blo<ck\red'
PASS fenced frame mode=opaque-ads dangling-markup URL with 'blo<ck	ed'
PASS fenced frame mode=default dangling-markup URL with 'blo
ck<ed'
PASS fenced frame mode=default dangling-markup URL with 'blo\rck<ed'
PASS fenced frame mode=default dangling-markup URL with 'blo	ck<ed'
PASS fenced frame mode=default dangling-markup URL with 'blo<ck
ed'
PASS fenced frame mode=default dangling-markup URL with 'blo<ck\red'
PASS fenced frame mode=default dangling-markup URL with 'blo<ck	ed'
FAIL fenced frame opaque URN => https: URL with dangling markup 'blo
ck<ed' assert_equals: expected "NOT LOADED" but got "https://web-platform.test:8444/wpt_internal/fenced_frame/resources/report-url.html?blo<cked="
FAIL fenced frame opaque URN => https: URL with dangling markup 'blo\rck<ed' assert_equals: expected "NOT LOADED" but got "https://web-platform.test:8444/wpt_internal/fenced_frame/resources/report-url.html?blo<cked="
FAIL fenced frame opaque URN => https: URL with dangling markup 'blo	ck<ed' assert_equals: expected "NOT LOADED" but got "https://web-platform.test:8444/wpt_internal/fenced_frame/resources/report-url.html?blo<cked="
FAIL fenced frame opaque URN => https: URL with dangling markup 'blo<ck
ed' assert_equals: expected "NOT LOADED" but got "https://web-platform.test:8444/wpt_internal/fenced_frame/resources/report-url.html?blo<cked="
FAIL fenced frame opaque URN => https: URL with dangling markup 'blo<ck\red' assert_equals: expected "NOT LOADED" but got "https://web-platform.test:8444/wpt_internal/fenced_frame/resources/report-url.html?blo<cked="
FAIL fenced frame opaque URN => https: URL with dangling markup 'blo<ck	ed' assert_equals: expected "NOT LOADED" but got "https://web-platform.test:8444/wpt_internal/fenced_frame/resources/report-url.html?blo<cked="
Harness: the test ran to completion.
